Fix ChunkDecoder premature completion when back-end closes connection
ChunckDecoder completes when the  endOfStream property is set to true. When there is SSLIOSession involved this property sets to tru soon after the in put channel reaches it's end of stream. But the SSLIOSession's inEncrypted buffer can be half consumed at this state. So done this fix to avoid setting endOfStream true until th inEncrypted buffer is Fully consumed.

 Fixes: wso2/wso2-httpcore-nio#49